requirements.txt
config==0.5.0.post0
cycler==0.10.0
Cython==0.29.24
Django==2.0.7
django-cors-headers==2.4.0
django-rest-framework==0.1.0
django-uwsgi==0.2.2
djangorestframework==3.8.2
imageio==2.9.0
joblib==1.0.1
kiwisolver==1.3.1
lmdb==1.2.1
lxml==4.6.3
matplotlib==3.3.4
networkx==2.5.1
numpy==1.19.5
opencv-python==4.5.3.56
pandas==1.1.5
Pillow==8.3.1
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2021.1
PyWavelets==1.1.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
threadpoolctl==2.2.0
tifffile==2020.9.3
typing-extensions==3.10.0.0
uwsgi-tasks==0.7.3
pytesseract==0.3.8
uwsgi==2.0.20
psycopg2-binary==2.9.1
sklearn==0.0
scikit-image==0.18.3
PyPDF2==1.26.0
pdf2image==1.16.0
pdfminer.six==20211012
bs4==0.0.1
fitz==0.0.1.dev2
pymupdf==1.19.2
